Update for 1.6.x, Use Bukkit and gson instead of craftbukkit, rewrite Essentials integration

+++ src/com/massivecraft/factions/listeners/FactionsEntityListener.java
@@ -243,7 +243,7 @@ public class FactionsEntityListener implements Listener
 	{
 		if (event.isCancelled()) return;
 		
-		EntityDamageByEntityEvent sub = new EntityDamageByEntityEvent(event.getCombuster(), event.getEntity(), EntityDamageEvent.DamageCause.FIRE, 0);
+		EntityDamageByEntityEvent sub = new EntityDamageByEntityEvent(event.getCombuster(), event.getEntity(), EntityDamageEvent.DamageCause.FIRE, 0.0D);
 		if ( ! this.canDamagerHurtDamagee(sub, false))
 			event.setCancelled(true);
 		sub = null;
@@ -278,7 +278,7 @@ public class FactionsEntityListener implements Listener
 		while (iter.hasNext())
 		{
 			LivingEntity target = iter.next();
-			EntityDamageByEntityEvent sub = new EntityDamageByEntityEvent(thrower, target, EntityDamageEvent.DamageCause.CUSTOM, 0);
+			EntityDamageByEntityEvent sub = new EntityDamageByEntityEvent(thrower, target, EntityDamageEvent.DamageCause.CUSTOM, 0.0D);
 			if ( ! this.canDamagerHurtDamagee(sub, true))
 				event.setIntensity(target, 0.0);  // affected entity list doesn't accept modification (so no iter.remove()), but this works
 			sub = null;
@@ -307,7 +307,7 @@ public class FactionsEntityListener implements Listener
 	{
 		Entity damager = sub.getDamager();
 		Entity damagee = sub.getEntity();
-		int damage = sub.getDamage();
+		double damage = sub.getDamage();
 		
 		if ( ! (damagee instanceof Player))
 			return true;
